fontend/requirements.txt
gradio==3.0.18
requests
google-generativeai==0.7.0 
python-dotenv
chromadb==0.4.24
chroma-hnswlib==0.7.3
langchain 
langchain-community
langchain-community
langchain_google_genai 
crewai 